backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.1
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
httpx==0.25.2
openai==1.3.8
google-generativeai==0.3.0
anthropic==0.7.8
firebase-admin==6.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
